Ensure mode file can be bundled
This fixes #27

The problem was that `build.sh` attempts to copy the generated mode file into the `src` folder in order to bundle it. If that folder doesn't exist, this step fails and the subsequent attempt to bundle things fails, too. Creating this file ensures the `src` folder will always be present.